pyproject.toml
[project]
name = "kembelearnbackend"
package-mode = false
version = "0.1.0"
description = ""
authors = [
    {name = "meligavincenttheophane",email = "meligavincenttheophane@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "llama-index (>=0.12.25,<0.13.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "llama-index-llms-huggingface-api (>=0.4.1,<0.5.0)",
    "llama-index-embeddings-huggingface (>=0.5.2,<0.6.0)",
    "llama-index-tools-google (>=0.3.0,<0.4.0)",
    "llama-index-utils-workflow (>=0.3.0,<0.4.0)",
    "gradio (>=5.23.3,<6.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain-mistralai (>=0.2.10,<0.3.0)",
    "groq (>=0.20.0,<0.21.0)",
    "spacy[lookups,transformers] (>=3.8.4,<4.0.0)",
    "pip (>=25.0.1,<26.0.0)",
    "setuptools (>=78.1.0,<79.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "catalogue (>=2.0.10,<3.0.0)",
    "soundfile (>=0.13.1,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"
alembic = "^1.15.1"

[tool.poetry.scripts]
download-spacy-models = "scripts.download_models:main"

